진짜 최종

+++ src/main/java/com/example/demo/src/review/model/PostReviewReq.java
@@ -5,6 +5,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+
 import java.util.List;
 
 @Getter
@@ -15,7 +16,7 @@ public class PostReviewReq {
     // 유저 id, 별점, 리뷰내용, 리뷰이미지 url
     private int userIdx;
 //    @NotNull(message = "별점을 입력해주세요.")
-    private Long star;
+    private float star;
     private String content;
     private List<String> reviewImgUrl;
